New Text Component (https://github.com/enso-org/ide/pull/132)
Refactored TextComponent to use the sprites and glyph system, and renamed it to TextField. Also do much refactoring stuff and add missing documentation.

Original commit: https://github.com/enso-org/ide/commit/49efefad2c87073a765a0eca42c647c67c299b85

+++ gui/lib/core/src/display/shape/text/glyph/font.rs
@@ -2,8 +2,9 @@
 
 use crate::prelude::*;
 
-use crate::display::shape::glyph::msdf::{MsdfTexture, convert_msdf_translation};
-use crate::display::shape::glyph::msdf::x_distance_from_msdf_value;
+use crate::display::shape::text::glyph::msdf::MsdfTexture;
+use crate::display::shape::text::glyph::msdf::convert_msdf_translation;
+use crate::display::shape::text::glyph::msdf::x_distance_from_msdf_value;
 
 use basegl_core_msdf_sys as msdf_sys;
 use basegl_core_embedded_fonts::EmbeddedFonts;
@@ -40,7 +41,7 @@ pub struct GlyphRenderInfo {
     /// A required scale of the _base square_ see structure documentation for details.
     pub scale: Vector2<f32>,
     /// An advance. Advance is the distance between two successive pen positions for specific glyph.
-    pub advance: f64
+    pub advance: f32
 }
 
 /// A single font data used for rendering
@@ -57,7 +58,7 @@ pub struct FontRenderInfo {
     msdf_sys_font : msdf_sys::Font,
     msdf_texture  : MsdfTexture,
     glyphs        : HashMap<char,GlyphRenderInfo>,
-    kerning       : HashMap<(char,char),f64>
+    kerning       : HashMap<(char,char),f32>
 }
 
 impl FontRenderInfo {
@@ -130,7 +131,7 @@ impl FontRenderInfo {
     }
 
     /// Get kerning between two characters
-    pub fn get_kerning(&mut self, left : char, right : char) -> f64 {
+    pub fn get_kerning(&mut self, left : char, right : char) -> f32 {
         match self.kerning.entry((left,right)) {
             Occupied(entry) => *entry.get(),
             Vacant(entry)   => {
@@ -174,8 +175,8 @@ impl FontRenderInfo {
     }
 
     #[cfg(test)]
-    pub fn mock_kerning_info(&mut self, l : char, r : char, value : f64) {
-        self.kerning.insert((l, r),value);
+    pub fn mock_kerning_info(&mut self, l : char, r : char, value : f32) {
+        self.kerning.insert((l,r),value);
     }
 }
 
@@ -237,7 +238,7 @@ impl Default for FontRegistry {
 #[cfg(test)]
 mod tests {
     use super::*;
-    use crate::display::shape::glyph::msdf::MsdfTexture;
+    use crate::display::shape::text::glyph::msdf::MsdfTexture;
 
     use basegl_core_msdf_sys as msdf_sys;
     use basegl_core_embedded_fonts::EmbeddedFonts;
